# backend/domain/camera-service.ts
import type { AuthUser } from './auth-service';

export type CameraStatus = 'Operativa' | 'En mantenimiento' | 'Fuera de servicio';

export interface CameraOperationsUpdate {
  status: CameraStatus;
  lastMaintenanceOn: string;
  nextMaintenanceOn: string;
  responsibleArea: string;
  notes: string;
}

export interface CameraUpdateCommand extends CameraOperationsUpdate {
  expectedVersion?: number;
}

export interface CameraRecord extends CameraOperationsUpdate {
  id: string;
  version: number;
  assetCode: string;
  name: string;
  panoramaId: string;
  location: string;
  brand: string;
  model: string;
  type: '360°' | 'Fija';
  yaw: number;
  pitch: number;
  installedOn: string;
  coverage: string;
  recordingMode: string;
  retention: string;
}

export interface CameraUpdateResult {
  camera: CameraRecord;
  changedFields: Array<keyof CameraOperationsUpdate>;
}

export interface CameraService {
  listCameras(): Promise<CameraRecord[]>;
  getCamera(cameraId: string): Promise<CameraRecord | null>;
  updateOperations(
    cameraId: string,
    actor: AuthUser,
    update: CameraUpdateCommand,
  ): Promise<CameraUpdateResult | null>;
}

export class CameraActorNotAllowedError extends Error {}

export class CameraVersionConflictError extends Error {}