# documentacion/openapi/orbinodo-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OrbiNodo Enterprise API",
    "version": "1.0.0",
    "description": "Contrato versionado de la API Fastify de OrbiNodo. Los datos CCTV son ficticios."
  },
  "servers": [
    { "url": "http://127.0.0.1:3001", "description": "Desarrollo local" },
    { "url": "https://orbinodo-demo.vercel.app", "description": "Produccion Vercel" }
  ],
...
  "paths": {
...
    "/api/cameras": {
      "get": {
        "tags": ["Camaras"], "summary": "Lista las camaras", "security": [{ "cookieAuth": [] }], "x-orbinodo-roles": ["programmer", "manager", "engineer1", "engineer2"],
        "responses": { "200": { "description": "Camaras disponibles", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CameraListResponse" } } } }, "401": { "$ref": "#/components/responses/SessionRequired" }, "500": { "$ref": "#/components/responses/InternalError" }, "426": { "$ref": "#/components/responses/HttpsRequired" } }
      },
      "head": { "tags": ["Camaras"], "summary": "Comprueba listado sin cuerpo", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "Disponible" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "500": { "$ref": "#/components/responses/InternalError" }, "426": { "$ref": "#/components/responses/HttpsRequired" } } }
    },
    "/api/cameras/{id}": {
      "get": {
        "tags": ["Camaras"], "summary": "Consulta una camara individual", "security": [{ "cookieAuth": [] }], "x-orbinodo-roles": ["programmer", "manager", "engineer1", "engineer2"],
        "parameters": [{ "$ref": "#/components/parameters/CameraId" }],
        "responses": { "200": { "description": "Camara encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CameraResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "404": { "$ref": "#/components/responses/CameraNotFound" }, "500": { "$ref": "#/components/responses/InternalError" }, "426": { "$ref": "#/components/responses/HttpsRequired" } }
      },
      "head": { "tags": ["Camaras"], "summary": "Comprueba una camara sin cuerpo", "security": [{ "cookieAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/CameraId" }], "responses": { "200": { "description": "Camara encontrada" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "404": { "$ref": "#/components/responses/CameraNotFound" }, "500": { "$ref": "#/components/responses/InternalError" }, "426": { "$ref": "#/components/responses/HttpsRequired" } } }
    },
    "/api/cameras/{id}/operations": {
      "patch": {
        "tags": ["Camaras"], "summary": "Actualiza campos operacionales autorizados", "security": [{ "cookieAuth": [] }], "x-orbinodo-roles": ["engineer1", "engineer2"],
        "parameters": [{ "$ref": "#/components/parameters/CameraId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CameraOperationsRequest" }, "example": { "status": "En mantenimiento", "lastMaintenanceOn": "2026-02-01", "nextMaintenanceOn": "2026-07-01", "responsibleArea": "Mantenimiento", "notes": "Revision programada", "expectedVersion": 0 } } } },
        "responses": { "200": { "description": "Actualizacion aplicada y auditada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CameraUpdateResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "403": { "$ref": "#/components/responses/ForbiddenRole" }, "404": { "$ref": "#/components/responses/CameraNotFound" }, "409": { "$ref": "#/components/responses/ResourceConflict" }, "413": { "$ref": "#/components/responses/PayloadTooLarge" }, "415": { "$ref": "#/components/responses/UnsupportedMediaType" }, "500": { "$ref": "#/components/responses/InternalError" }, "426": { "$ref": "#/components/responses/HttpsRequired" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "orbinodo_session", "description": "Cookie HttpOnly creada por POST /api/auth/login." } },
    "parameters": {
      "CameraId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^camera-[0-9]{2}$" }, "example": "camera-01" },
...
    },
    "schemas": {
      "ApiError": { "type": "object", "additionalProperties": false, "required": ["error", "message"], "properties": { "error": { "type": "string", "enum": ["VALIDATION_ERROR", "INVALID_JSON", "PAYLOAD_TOO_LARGE", "UNSUPPORTED_MEDIA_TYPE", "HTTPS_REQUIRED", "CORS_ORIGIN_FORBIDDEN", "METHOD_NOT_ALLOWED", "API_RATE_LIMITED", "INVALID_CREDENTIALS", "LOGIN_RATE_LIMITED", "SESSION_REQUIRED", "FORBIDDEN_ROLE", "CAMERA_NOT_FOUND", "ROUTE_NOT_FOUND", "RESOURCE_CONFLICT", "INTERNAL_ERROR", "DATABASE_UNAVAILABLE"] }, "message": { "type": "string" } }, "example": { "error": "CAMERA_NOT_FOUND", "message": "La camara solicitada no existe." } },
...
      "CameraStatus": { "type": "string", "enum": ["Operativa", "En mantenimiento", "Fuera de servicio"] },
      "Camera": { "type": "object", "additionalProperties": false, "required": ["id", "version", "assetCode", "name", "panoramaId", "location", "brand", "model", "type", "yaw", "pitch", "status", "installedOn", "lastMaintenanceOn", "nextMaintenanceOn", "coverage", "recordingMode", "retention", "responsibleArea", "notes"], "properties": { "id": { "type": "string", "pattern": "^camera-[0-9]{2}$" }, "version": { "type": "integer", "minimum": 0 }, "assetCode": { "type": "string" }, "name": { "type": "string" }, "panoramaId": { "type": "string" }, "location": { "type": "string" }, "brand": { "type": "string" }, "model": { "type": "string" }, "type": { "type": "string", "enum": ["360?", "Fija"] }, "yaw": { "type": "number" }, "pitch": { "type": "number" }, "status": { "$ref": "#/components/schemas/CameraStatus" }, "installedOn": { "type": "string", "format": "date" }, "lastMaintenanceOn": { "type": "string", "format": "date" }, "nextMaintenanceOn": { "type": "string", "format": "date" }, "coverage": { "type": "string" }, "recordingMode": { "type": "string" }, "retention": { "type": "string" }, "responsibleArea": { "type": "string" }, "notes": { "type": "string" } } },
      "CameraOperationsRequest": { "type": "object", "additionalProperties": false, "required": ["status", "lastMaintenanceOn", "nextMaintenanceOn", "responsibleArea", "notes", "expectedVersion"], "properties": { "status": { "$ref": "#/components/schemas/CameraStatus" }, "lastMaintenanceOn": { "type": "string", "format": "date" }, "nextMaintenanceOn": { "type": "string", "format": "date" }, "responsibleArea": { "type": "string", "minLength": 1, "maxLength": 100 }, "notes": { "type": "string", "minLength": 1, "maxLength": 500 }, "expectedVersion": { "type": "integer", "minimum": 0 } } },
      "CameraListResponse": { "type": "object", "additionalProperties": false, "required": ["cameras"], "properties": { "cameras": { "type": "array", "items": { "$ref": "#/components/schemas/Camera" } } } },
      "CameraResponse": { "type": "object", "additionalProperties": false, "required": ["camera"], "properties": { "camera": { "$ref": "#/components/schemas/Camera" } } },
      "CameraUpdateResponse": { "type": "object", "additionalProperties": false, "required": ["camera", "changedFields"], "properties": { "camera": { "$ref": "#/components/schemas/Camera" }, "changedFields": { "type": "array", "items": { "type": "string", "enum": ["status", "lastMaintenanceOn", "nextMaintenanceOn", "responsibleArea", "notes"] } } } },
...
    },
    "responses": {
      "ValidationError": { "description": "Datos, parametros o filtros invalidos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" }, "example": { "error": "VALIDATION_ERROR", "message": "Los datos de la solicitud no son validos." } } } },
...
      "SessionRequired": { "description": "Sesion ausente, vencida o revocada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
      "ForbiddenRole": { "description": "Sesion valida sin permiso", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" }, "example": { "error": "FORBIDDEN_ROLE", "message": "Este perfil no puede editar c?maras." } } } },
      "CameraNotFound": { "description": "Camara inexistente", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
      "ResourceConflict": { "description": "Version de camara desactualizada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" }, "example": { "error": "RESOURCE_CONFLICT", "message": "La c?mara fue modificada por otro usuario." } } } },
      "PayloadTooLarge": { "description": "Cuerpo superior a 16 KiB", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
      "UnsupportedMediaType": { "description": "Tipo de contenido no compatible", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
...
      "HttpsRequired": { "description": "HTTPS requerido en produccion", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
...
      "InternalError": { "description": "Error interno sin detalles sensibles", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
    }
  }
}